fix: client not sending move to server
fixed a bug where when 2 players where connected and one moves, no data is sent to the server.
This was happening because the server was thinking that the json data received for the target position was already a position object,
and was trying to access the string form of the object name eg "g6" by saying targetsquare.name.
this would have worked if it was a position object, but the json data was being sent as a string already.
so i just removed the .name

as of now the second player is receiving the updated game state after the first move, but when the second player moves there is no update on screen.