test: added backwards compatibility for tests
previous commit fixed functionality but failed a previous test.
instead of refactoring all my tests to change position types to strings for position name,
I just added a small changd to the code, which keeps functionality the same.
It doesnt really matter if the client sends position objects or strings, this was just to
avoid changing all tests.